requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
beautifulsoup4==4.10.0
httpx==0.22.0
lxml==4.7.1
pathos==0.2.8
Pillow==9.0.1
pypinyin==0.46.0
PyYAML==6.0
requests==2.27.1
ujson==5.1.0
prettytable==3.2.0
GitPython==3.1.27